feat: added basic prometheus instrumentation
